+++ include/CppLinuxSerial/SerialPort.hpp
@@ -32,7 +32,6 @@ namespace mn {
         };
 
         enum class State {
-            UNCONFIGURED,
             CLOSED,
             OPEN
         };
@@ -41,27 +40,32 @@ namespace mn {
         class SerialPort {
 
         public:
-            /// \brief		Default constructor.
+            /// \brief		Default constructor. You must specify at least the device before calling Open().
             SerialPort();
 
-            /// \brief		Constructor that sets up serial port with all required parameters.
+            /// \brief		Constructor that sets up serial port with the basic (required) parameters.
             SerialPort(const std::string &device, BaudRate baudRate);
 
             //! @brief		Destructor. Closes serial port if still open.
             virtual ~SerialPort();
 
-            //! @brief		Sets the file path to use for communications. The file path must be set before Open() is called, otherwise Open() will return an error.
+            /// \brief		Sets the device to use for serial port communications.
+            /// \details    Method can be called when serial port is in any state.
             void SetDevice(const std::string &device);
 
             void SetBaudRate(BaudRate baudRate);
 
-            //! @brief		Controls what happens when Read() is called.
-            //! @param		numOfCharToWait		Minimum number of characters to wait for before returning. Set to 0 for non-blocking mode.
-            void SetNumCharsToWait(uint32_t numCharsToWait);
+            /// \brief      Sets the read timeout (in milliseconds)/blocking mode.
+            /// \details    Only call when state != OPEN. This method manupulates VMIN and VTIME.
+            /// \param      timeout_ms  Set to -1 to infinite timeout, 0 to return immediately with any data (non
+            ///             blocking, or >0 to wait for data for a specified number of milliseconds). Timeout will
+            ///             be rounded to the nearest 100ms (a Linux API restriction). Maximum value limited to
+            ///             25500ms (another Linux API restriction).
+            void SetTimeout(int32_t timeout_ms);
 
-            //! @brief		Enables/disables echo.
-            //! param		echoOn		Pass in true to enable echo, false to disable echo.
-            void EnableEcho(bool echoOn);
+            /// \brief		Enables/disables echo.
+            /// \param		value		Pass in true to enable echo, false to disable echo.
+            void SetEcho(bool value);
 
             //! @brief		Opens the COM port for use.
             //! @throws		{std::runtime_error} if filename has not been set.
@@ -69,41 +73,52 @@ namespace mn {
             //! @note		Must call this before you can configure the COM port.
             void Open();
 
-            /// \brief		Configures the tty device as a serial port.
-            /// \warning    Device must be open (valid file descriptor) when this is called.
-            void ConfigureTermios();
-
-            //! @brief		Closes the COM port.
+            /// \brief		Closes the COM port.
             void Close();
 
-            //! @brief		Sends a message over the com port.
-            //! @param		str		Reference to an string containing the characters to write to the COM port.
-            //! @throws		{std::runtime_error} if filename has not been set.
-            //!				{std::system_error} if system write() operation fails.
+            /// \brief		Sends a message over the com port.
+            /// \param		data		The data that will be written to the COM port.
+            /// \throws		CppLinuxSerial::Exception if state != OPEN.
             void Write(const std::string& data);
 
-            //! @brief		Use to read from the COM port.
-            //! @param		str		Reference to a string that the read characters from the COM port will be saved to.
-            //! @throws		{std::runtime_error} if filename has not been set.
-            //!				{std::system_error} if system read() operation fails.
+            /// \brief		Use to read from the COM port.
+            /// \param		data		The object the read characters from the COM port will be saved to.
+            /// \param      wait_ms     The amount of time to wait for data. Set to 0 for non-blocking mode. Set to -1
+            ///                 to wait indefinitely for new data.
+            /// \throws		CppLinuxSerial::Exception if state != OPEN.
             void Read(std::string& data);
 
         private:
 
+            /// \brief		Configures the tty device as a serial port.
+            /// \warning    Device must be open (valid file descriptor) when this is called.
+            void ConfigureTermios();
+
+            void SetTermios(termios myTermios);
+
             /// \brief      Keeps track of the serial port's state.
             State state_;
 
+            /// \brief      The file path to the serial port device (e.g. "/dev/ttyUSB0").
             std::string device_;
 
+            /// \brief      The current baud rate.
             BaudRate baudRate_;
 
-            //! @brief		The file descriptor for the open file. This gets written to when Open() is called.
+            /// \brief		The file descriptor for the open file. This gets written to when Open() is called.
             int fileDesc_;
 
-            //! @brief		Returns a populated termios structure for the passed in file descriptor.
+            bool echo_;
+
+            int32_t timeout_ms_;
+
+            /// \brief		Returns a populated termios structure for the passed in file descriptor.
             termios GetTermios();
 
-            void SetTermios(termios myTermios);
+            static constexpr BaudRate defaultBaudRate_ = BaudRate::B_57600;
+            static constexpr int32_t defaultTimeout_ms_ = -1;
+
+
         };
 
     } // namespace CppLinuxSerial