+++ src/SerialPort.cpp
@@ -2,11 +2,12 @@
 //! @file 			SerialPort.cpp
 //! @author 		Geoffrey Hunter <gbmhunter@gmail.com> (www.mbedded.ninja)
 //! @created		2014-01-07
-//! @last-modified 	2017-11-23
+//! @last-modified 	2017-11-27
 //! @brief			The main serial port class.
 //! @details
 //!					See README.rst in repo root dir for more info.
 
+// System includes
 #include <iostream>
 #include <sstream>
 #include <stdio.h>   	// Standard input/output definitions
@@ -17,17 +18,23 @@
 #include <termios.h> 	// POSIX terminal control definitions (struct termios)
 #include <system_error>	// For throwing std::system_error
 
+// User includes
+#include "CppLinuxSerial/Exception.hpp"
 #include "CppLinuxSerial/SerialPort.hpp"
 
 namespace mn {
 namespace CppLinuxSerial {
 
 	SerialPort::SerialPort() {
+        echo_ = false;
+        timeout_ms_ = defaultTimeout_ms_;
+        baudRate_ = defaultBaudRate_;
 	}
 
-	SerialPort::SerialPort(const std::string& device, BaudRate baudRate) {
+	SerialPort::SerialPort(const std::string& device, BaudRate baudRate) :
+            SerialPort() {
 		device_ = device;
-		baudRate_ = baudRate;
+        baudRate_ = baudRate;
 	}
 
 	SerialPort::~SerialPort() {
@@ -35,19 +42,22 @@ namespace CppLinuxSerial {
             Close();
         } catch(...) {
             // We can't do anything about this!
+            // But we don't want to throw within destructor, so swallow
         }
 	}
 
-	void SerialPort::SetDevice(const std::string& device)
-	{		
+	void SerialPort::SetDevice(const std::string& device) {
 		device_ = device;
+        if(state_ == State::OPEN)
+
+
         ConfigureTermios();
 	}
 
-	void SerialPort::SetBaudRate(BaudRate baudRate)
-	{
+	void SerialPort::SetBaudRate(BaudRate baudRate)	{
 		baudRate_ = baudRate;
-        ConfigureTermios();
+        if(state_ == State::OPEN)
+            ConfigureTermios();
 	}
 
 	void SerialPort::Open()
@@ -59,7 +69,7 @@ namespace CppLinuxSerial {
 			//this->sp->PrintError(SmartPrint::Ss() << "Attempted to open file when file path has not been assigned to.");
 			//return false;
 
-			throw std::runtime_error("Attempted to open file when file path has not been assigned to.");
+			THROW_EXCEPT("Attempted to open file when file path has not been assigned to.");
 		}
 
 		// Attempt to open file
@@ -75,25 +85,18 @@ namespace CppLinuxSerial {
 		    //this->sp->PrintError(SmartPrint::Ss() << "Unable to open " << this->filePath << " - " << strerror(errno));
 		    //return false;
 
-		    throw std::runtime_error("Could not open device " + device_ + ". Is the device name correct and do you have read/write permission?");
+		    THROW_EXCEPT("Could not open device " + device_ + ". Is the device name correct and do you have read/write permission?");
 		}
 
         ConfigureTermios();
 
 		std::cout << "COM port opened successfully." << std::endl;
-
-		// If code reaches here, open and config must of been successful
-
+        state_ = State::OPEN;
 	}
 
-	void SerialPort::EnableEcho(bool echoOn)
-	{
-		termios settings = this->GetTermios();
-		settings.c_lflag = echoOn
-					   ? (settings.c_lflag |   ECHO )
-					   : (settings.c_lflag & ~(ECHO));
-		//tcsetattr( STDIN_FILENO, TCSANOW, &settings );
-		this->SetTermios(settings);
+	void SerialPort::SetEcho(bool value) {
+        echo_ = value;
+        ConfigureTermios();
 	}
 
 	void SerialPort::ConfigureTermios()
@@ -147,14 +150,29 @@ namespace CppLinuxSerial {
 
 		//================= CONTROL CHARACTERS (.c_cc[]) ==================//
 
-		// c_cc[WMIN] sets the number of characters to block (wait) for when read() is called.
-		// Set to 0 if you don't want read to block. Only meaningful when port set to non-canonical mode
-		//tty.c_cc[VMIN]      =   1;
-		SetNumCharsToWait(1);
-
 		// c_cc[VTIME] sets the inter-character timer, in units of 0.1s.
 		// Only meaningful when port is set to non-canonical mode
-		tty.c_cc[VTIME]     =   5;          // 0.5 seconds read timeout
+        // VMIN = 0, VTIME = 0: No blocking, return immediately with what is available
+        // VMIN > 0, VTIME = 0: read() waits for VMIN bytes, could block indefinitely
+        // VMIN = 0, VTIME > 0: Block until any amount of data is available, OR timeout occurs
+        // VMIN > 0, VTIME > 0: Block until either VMIN characters have been received, or VTIME
+        //                      after first character has elapsed
+        // c_cc[WMIN] sets the number of characters to block (wait) for when read() is called.
+        // Set to 0 if you don't want read to block. Only meaningful when port set to non-canonical mode
+
+        if(timeout_ms_ == -1) {
+            // Always wait for at least one byte, this could
+            // block indefinitely
+            tty.c_cc[VTIME] = 0;
+            tty.c_cc[VMIN] = 1;
+        } else if(timeout_ms_ == 0) {
+            // Setting both to 0 will give a non-blocking read
+            tty.c_cc[VTIME] = 0;
+            tty.c_cc[VMIN] = 0;
+        } else if(timeout_ms_ > 0) {
+            tty.c_cc[VTIME] = (cc_t)(timeout_ms_/100);    // 0.5 seconds read timeout
+            tty.c_cc[VMIN] = 0;
+        }
 
 
 		//======================== (.c_iflag) ====================//
@@ -162,15 +180,19 @@ namespace CppLinuxSerial {
 		tty.c_iflag     &= ~(IXON | IXOFF | IXANY);			// Turn off s/w flow ctrl
 		tty.c_iflag 	&= ~(IGNBRK|BRKINT|PARMRK|ISTRIP|INLCR|IGNCR|ICRNL);
 
+
+
 		//=========================== LOCAL MODES (c_lflag) =======================//
 
 		// Canonical input is when read waits for EOL or EOF characters before returning. In non-canonical mode, the rate at which
 		// read() returns is instead controlled by c_cc[VMIN] and c_cc[VTIME]
 		tty.c_lflag		&= ~ICANON;								// Turn off canonical input, which is suitable for pass-through
-		tty.c_lflag		&= ~ECHO;								// Turn off echo
+        echo_ ? (tty.c_lflag | ECHO ) : (tty.c_lflag & ~(ECHO));	// Configure echo depending on echo_ boolean
 		tty.c_lflag		&= ~ECHOE;								// Turn off echo erase (echo erase only relevant if canonical input is active)
 		tty.c_lflag		&= ~ECHONL;								//
 		tty.c_lflag		&= ~ISIG;								// Disables recognition of INTR (interrupt), QUIT and SUSP (suspend) characters
+
+
 
 		// Try and use raw function call
 		//cfmakeraw(&tty);
@@ -189,38 +211,21 @@ namespace CppLinuxSerial {
 		}*/
 	}
 
-	void SerialPort::SetNumCharsToWait(uint32_t numCharsToWait) {
-		// Get current termios struct
-		termios myTermios = GetTermios();
-
-		// Save the number of characters to wait for
-		// to the control register
-		myTermios.c_cc[VMIN] = numCharsToWait;
-
-		// Save termios back
-		SetTermios(myTermios);
-	}
-
 	void SerialPort::Write(const std::string& data) {
-		if(fileDesc_ == 0) {
-			//this->sp->PrintError(SmartPrint::Ss() << );
-			//return false;
 
-			throw std::runtime_error("SendMsg called but file descriptor (fileDesc) was 0, indicating file has not been opened.");
+        if(state_ != State::OPEN)
+            THROW_EXCEPT(std::string() + __PRETTY_FUNCTION__ + " called but state != OPEN. Please call Open() first.");
+
+		if(fileDesc_ < 0) {
+			THROW_EXCEPT(std::string() + __PRETTY_FUNCTION__ + " called but file descriptor < 0, indicating file has not been opened.");
 		}
 
 		int writeResult = write(fileDesc_, data.c_str(), data.size());
 
 		// Check status
 		if (writeResult == -1) {
-			// Could not open COM port
-			//this->sp->PrintError(SmartPrint::Ss() << "Unable to write to \"" << this->filePath << "\" - " << strerror(errno));
-			//return false;
-
 			throw std::system_error(EFAULT, std::system_category());
 		}
-
-		// If code reaches here than write must of been successful
 	}
 
 	void SerialPort::Read(std::string& data)
@@ -228,7 +233,7 @@ namespace CppLinuxSerial {
 		if(fileDesc_ == 0) {
 			//this->sp->PrintError(SmartPrint::Ss() << "Read() was called but file descriptor (fileDesc) was 0, indicating file has not been opened.");
 			//return false;
-			throw std::runtime_error("Read() was called but file descriptor (fileDesc) was 0, indicating file has not been opened.");
+			THROW_EXCEPT("Read() was called but file descriptor (fileDesc) was 0, indicating file has not been opened.");
 		}
 
 		// Allocate memory for read buffer
@@ -236,7 +241,7 @@ namespace CppLinuxSerial {
 		memset (&buf, '\0', sizeof buf);
 
 		// Read from file
-		int n = read(fileDesc_, &buf, sizeof(buf));
+		ssize_t n = read(fileDesc_, &buf, sizeof(buf));
 
 		// Error Handling
 		if(n < 0) {
@@ -256,8 +261,7 @@ namespace CppLinuxSerial {
 		// If code reaches here, read must of been successful
 	}
 
-	termios SerialPort::GetTermios()
-	{
+	termios SerialPort::GetTermios() {
         if(fileDesc_ == -1)
             throw std::runtime_error("GetTermios() called but file descriptor was not valid.");
 
@@ -279,9 +283,9 @@ namespace CppLinuxSerial {
 	void SerialPort::SetTermios(termios myTermios)
 	{
 		// Flush port, then apply attributes
-		tcflush(this->fileDesc_, TCIFLUSH);
+		tcflush(fileDesc_, TCIFLUSH);
 
-		if(tcsetattr(this->fileDesc_, TCSANOW, &myTermios) != 0)
+		if(tcsetattr(fileDesc_, TCSANOW, &myTermios) != 0)
 		{
 			// Error occurred
 			std::cout << "Could not apply terminal attributes for \"" << device_ << "\" - " << strerror(errno) << std::endl;
@@ -296,12 +300,22 @@ namespace CppLinuxSerial {
         if(fileDesc_ != -1) {
             auto retVal = close(fileDesc_);
             if(retVal != 0)
-                throw std::runtime_error("Tried to close serial port " + device_ + ", but close() failed.");
+                THROW_EXCEPT("Tried to close serial port " + device_ + ", but close() failed.");
 
             fileDesc_ = -1;
         }
 
         state_ = State::CLOSED;
+    }
+
+    void SerialPort::SetTimeout(int32_t timeout_ms) {
+        if(timeout_ms < -1)
+            THROW_EXCEPT(std::string() + "timeout_ms provided to " + __PRETTY_FUNCTION__ + " was < -1, which is invalid.");
+        if(timeout_ms > 25500)
+            THROW_EXCEPT(std::string() + "timeout_ms provided to " + __PRETTY_FUNCTION__ + " was > 25500, which is invalid.");
+        if(state_ == State::OPEN)
+            THROW_EXCEPT(std::string() + __PRETTY_FUNCTION__ + " called while state == OPEN.");
+        timeout_ms_ = timeout_ms;
     }
 
 } // namespace CppLinuxSerial